requirements.txt
Flask==2.3.3
python-whois==0.8.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3 